refactored and added the base of the navbar
