+++ handlingo/app/(authentication)/forgot-password/page.tsx
@@ -9,7 +9,7 @@
 // export default async function ForgotPassword(props: {
 //   searchParams: Promise<Message>;
 // }) {
-//   const searchParams = await props.searchParams;
+    //   const searchParams = await props.searchParams;
 //   return (
 //     <>
 //       <form className="flex-1 flex flex-col w-full gap-2 text-foreground [&>input]:mb-6 min-w-64 max-w-64 mx-auto">
