+++ handlingo/app/(authentication)/sign-in/page.tsx
@@ -15,28 +15,8 @@ export default async function SignIn(props: { searchParams: Promise<Message> }) 
   const searchParams = await props.searchParams;
 
   return (
-    <div className="flex w-screen items-center justify-center min-h-screen relative">
-      {/* Background Image */}
-      <Image
-        src="/assets/login-background.png" // Use the imported image
-        alt="Login Background"
-        className="absolute top-0 left-0 w-full h-full object-full z-0"
-        width= {800}
-        height={800}
-      />
-
-      <form className="relative flex-1 flex flex-col min-w-48 max-w-96 z-10">
-      {/* Handlingo Title */}
-        <div className="flex justify-center">
-          <Image
-            src="/assets/handlingo-title.png"
-            alt="Handlingo Login Title"
-            className="mb-10 w-96 h-auto"
-            width={600} // Adjust the width as needed
-            height={150}  // Adjust the height as needed
-          />
-        </div>
-
+      
+      <form className="relative flex-1 flex flex-col min-w-48 max-w-96 w-3/12 z-10">
       {/* Log In Box */}
       <div className="flex flex-col items-center space-y-4 p-6 bg-lightBlue rounded-lg">
         
@@ -46,7 +26,7 @@ export default async function SignIn(props: { searchParams: Promise<Message> }) 
             Log in
           </h1>
 
-          <Link className="w-1/2 text-sm font-semibold font-fira text-gray mb-4 pb-2 border-b-2 border-lightBlue max-w-max" href="/auth-pages/sign-up">
+          <Link className="w-1/2 text-sm font-semibold font-fira text-gray mb-4 pb-2 border-b-2 border-lightBlue max-w-max" href="/sign-up">
             Sign Up
           </Link>
         </div>
@@ -85,7 +65,6 @@ export default async function SignIn(props: { searchParams: Promise<Message> }) 
       <FormMessage message={searchParams} />
       </div> {/* end blue box container */}
     </form> 
-  </div>           
   );
 }
