+++ handlingo/app/(main-pages)/layout.tsx
@@ -1,0 +1,15 @@
+import Navbar from '@/components/ui/navbar';
+
+export default async function Layout({
+    children,
+  }: {
+    children: React.ReactNode;
+  }) {
+    return (
+     <>
+      {/* Navbar */}
+      <Navbar />
+      {children}
+     </>
+    );
+  }
